+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/aggregate-chart-configuration.validation-schema.ts
@@ -20,8 +20,8 @@ import { CalendarStartDay } from 'twenty-shared/constants';
 
 import { AggregateOperations } from 'src/engine/api/graphql/graphql-query-runner/constants/aggregate-operations.constant';
 import { UUIDScalarType } from 'src/engine/api/graphql/workspace-schema-builder/graphql-types/scalars';
-import { RichTextV2BodyValidationSchema } from 'src/engine/metadata-modules/page-layout-widget/dtos/rich-text-v2-body.validation-schema';
 import { WidgetConfigurationType } from 'src/engine/metadata-modules/page-layout-widget/enums/widget-configuration-type.type';
+import { RichTextV2BodyValidationSchema } from 'src/engine/metadata-modules/page-layout-widget/schemas/rich-text-v2-body.validation-schema';
 import { ChartFilter } from 'src/engine/metadata-modules/page-layout-widget/types/chart-filter.type';
 import { PageLayoutWidgetConfigurationBase } from 'src/engine/metadata-modules/page-layout-widget/types/page-layout-widget-configurationt-base.type';
