+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/bar-chart-configuration.validation-schema.ts
