+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/calendar-configuration.validation-schema.ts
