+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/emails-configuration.validation-schema.ts
