+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/field-rich-text-configuration.validation-schema.ts
