+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/fields-configuration.validation-schema.ts
