+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/files-configuration.validation-schema.ts
