+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/gauge-chart-configuration.validation-schema.ts
