+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/inputs/create-page-layout-widget.input.ts
@@ -2,45 +2,49 @@ import { Field, InputType } from '@nestjs/graphql';
 
 import { Type } from 'class-transformer';
 import {
-  IsEnum,
-  IsObject,
-  IsOptional,
-  IsString,
-  IsUUID,
-  ValidateNested,
+    IsEnum,
+    IsNotEmpty,
+    IsObject,
+    IsOptional,
+    IsString,
+    IsUUID,
+    ValidateNested,
 } from 'class-validator';
 import { GraphQLJSON } from 'graphql-type-json';
 
 import { UUIDScalarType } from 'src/engine/api/graphql/workspace-schema-builder/graphql-types/scalars';
-import { GridPositionInput } from 'src/engine/metadata-modules/page-layout-widget/dtos/inputs/grid-position.input';
 import { WidgetType } from 'src/engine/metadata-modules/page-layout-widget/enums/widget-type.enum';
+import { GridPositionInput } from 'src/engine/metadata-modules/page-layout-widget/schemas/inputs/grid-position.input';
 import { AllPageLayoutWidgetConfiguration } from 'src/engine/metadata-modules/page-layout-widget/types/all-page-layout-widget-configuration.type';
 
 @InputType()
-export class UpdatePageLayoutWidgetInput {
-  @Field({ nullable: true })
-  @IsString()
-  @IsOptional()
-  title?: string;
+export class CreatePageLayoutWidgetInput {
+  @Field(() => UUIDScalarType, { nullable: false })
+  @IsUUID()
+  @IsNotEmpty()
+  pageLayoutTabId: string;
 
-  @Field(() => WidgetType, { nullable: true })
+  @Field({ nullable: false })
+  @IsString()
+  @IsNotEmpty()
+  title: string;
+
+  @Field(() => WidgetType, { nullable: false })
   @IsEnum(WidgetType)
-  @IsOptional()
-  type?: WidgetType;
+  type: WidgetType;
 
   @Field(() => UUIDScalarType, { nullable: true })
   @IsUUID()
   @IsOptional()
   objectMetadataId?: string | null;
 
-  @Field(() => GridPositionInput, { nullable: true })
+  @Field(() => GridPositionInput, { nullable: false })
   @ValidateNested()
   @Type(() => GridPositionInput)
-  @IsOptional()
-  gridPosition?: GridPositionInput;
+  gridPosition: GridPositionInput;
 
-  @Field(() => GraphQLJSON, { nullable: true })
+  @Field(() => GraphQLJSON, { nullable: false })
   @IsObject()
   @IsOptional()
-  configuration?: AllPageLayoutWidgetConfiguration;
+  configuration: AllPageLayoutWidgetConfiguration;
 }
