+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/inputs/delete-page-layout-widget.input.ts
