+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/inputs/destroy-page-layout-widget.input.ts
