+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/line-chart-configuration.validation-schema.ts
