+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/notes-configuration.validation-schema.ts
