+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/pie-chart-configuration.validation-schema.ts
