+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/ratio-aggregate-config.validation-schema.ts
