+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/rich-text-v2-body.validation-schema.ts
