+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/standalone-rich-text-configuration.validation-schema.ts
@@ -3,8 +3,8 @@ import { Field, ObjectType } from '@nestjs/graphql';
 import { Type } from 'class-transformer';
 import { IsIn, IsNotEmpty, ValidateNested } from 'class-validator';
 
-import { RichTextV2BodyValidationSchema } from 'src/engine/metadata-modules/page-layout-widget/dtos/rich-text-v2-body.validation-schema';
 import { WidgetConfigurationType } from 'src/engine/metadata-modules/page-layout-widget/enums/widget-configuration-type.type';
+import { RichTextV2BodyValidationSchema } from 'src/engine/metadata-modules/page-layout-widget/schemas/rich-text-v2-body.validation-schema';
 import { PageLayoutWidgetConfigurationBase } from 'src/engine/metadata-modules/page-layout-widget/types/page-layout-widget-configurationt-base.type';
 
 @ObjectType('StandaloneRichTextConfiguration')
