+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/timeline-configuration.validation-schema.ts
