+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/view-configuration.validation-schema.ts
