+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/widget-configuration.interface.ts
