+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/workflow-configuration.validation-schema.ts
