+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/schemas/workflow-version-configuration.validation-schema.ts
